pyproject.toml
[tool.poetry]
name = "anki-llm"
version = "0.1.0"
description = ""
authors = ["ciaran <ciaranregan123@gmai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ankipandas = "^0.3.14"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
